requirements.txt
deta==1.2.0
numpy==1.26.2
Requests==2.31.0
st_annotated_text==4.0.1
streamlit==1.29.0
streamlit_authenticator==0.2.3
streamlit_lottie==0.0.5
